% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addWord.R
\name{addWord}
\alias{addWord}
\title{Adds a single word or string}
\usage{
addWord(tree, string)
}
\arguments{
\item{tree}{an existing ternary search tree to add words to.}

\item{string}{a string of characters to be added to the tree.}
}
\value{
An object of class `list` and `tstTree`.
}
\description{
Adds a single word to an existing ternary search tree
}
\details{
The string of characters is added to the existing tree.
}
\examples{
fruitTree <- newTree(c("apple", "orange"))
fruitTree <- addWord(fruitTree, "lemon")
dimTree(fruitTree)
}
\seealso{
\code{\link{newTree}}
}

