% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/completeWord.R
\name{completeWord}
\alias{completeWord}
\title{Autocompletion of strings}
\usage{
completeWord(tree, string)
}
\arguments{
\item{tree}{an existing ternary search tree to search words in.}

\item{string}{a string of characters to be completed.}
}
\value{
A vector of class `character`.
}
\description{
Returns a character vector with the completed words
}
\details{
Searches recursively through the tree until all words starting with the 
specified string have been found, and stores them in a character vector.
}
\examples{
fruitTree <- newTree(c("apple", "orange","apricot","cherry"))
fruits.ap <- completeWord(fruitTree, "ap")
fruits.ap
}

