% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_by_call_rate.R
\name{dedup_by_callrate}
\alias{dedup_by_callrate}
\title{When a dataframe with characteristics from same inividuals (each replicate one row), for each individual, select the row with least ratio of NAs, leave each individual only one row.}
\usage{
dedup_by_callrate(df)
}
\arguments{
\item{df}{a dataframe with sample "ID" as its first column}
}
\value{
a data frame with 'ID' column de-duplicated by using ID of least NA rate
}
\description{
When a dataframe with characteristics from same inividuals (each replicate one row), for each individual, select the row with least ratio of NAs, leave each individual only one row.
}
\examples{
df example:
               ID purity ploidy
 TCGA-OR-A5J1-01A    0.9      2
 TCGA-OR-A5J2-01A   0.89    1.3
 TCGA-OR-A5J3-01A   0.93   1.27
 TCGA-OR-A5J4-01A   0.87    2.6
 TCGA-OR-A5J5-01A   0.93   2.79
}
