% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dimTree.R
\name{dimTree}
\alias{dimTree}
\title{Tree dimensions}
\usage{
dimTree(tree)
}
\arguments{
\item{tree}{an existing ternary search tree.}
}
\value{
A numeric vector with the dimensions of the tree.
}
\description{
Returns a numeric vector with the dimensions of the ternary search tree
}
\details{
The first number in the vector is the number of words in the ternary search tree 
and the second is the number of nodes (each node is a character) in the tree.
}
\examples{
fruitTree <- newTree(c("apple", "orange","apricot","cherry"))
dimTree(fruitTree)
}
\seealso{
\code{\link{newTree}}
}

