% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsg_filter_callrate.R
\name{dsg_filter_callrate}
\alias{dsg_filter_callrate}
\title{Filter dosaeg dataframe by call rate}
\usage{
dsg_filter_callrate(mydf, site_CR_threshold)
}
\arguments{
\item{mydf}{result of gather_dsg_DF, a dataframe of dosages, loci in the first column, and sampleID as colname}

\item{site_CR_threshold, }{anumber between 0-1, decimals}
}
\value{

}
\description{
Filter dosaeg dataframe by call rate
}
