% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_avrage_readlen.R
\name{fq_avrage_readlen}
\alias{fq_avrage_readlen}
\title{detect read length of fastq files, return the mean length of all the reads in given fastq files, compress status auto detected}
\usage{
fq_avrage_readlen(file_path)
}
\arguments{
\item{file_path}{fastq file, auto detect gz file}
}
\value{
mean read length
}
\description{
detect read length of fastq files, return the mean length of all the reads in given fastq files, compress status auto detected
}
