% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_dsg_DF.R
\name{gather_dsg_DF}
\alias{gather_dsg_DF}
\title{Gather dosage information to a dataframe from a directory containing gangSTR vcfs (wrapper of vcf2GT and GT2dsg)}
\usage{
gather_dsg_DF(mydir)
}
\arguments{
\item{mydir}{the directory of vcf files created by gangSTR}
}
\value{
a dataframe of dosages, loci in the first column, and sampleID as colname
}
\description{
Gather dosage information to a dataframe from a directory containing gangSTR vcfs (wrapper of vcf2GT and GT2dsg)
}
