% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correspond_sample_info.R
\name{get_correspond_sample_info}
\alias{get_correspond_sample_info}
\title{Function used to map GTEx v7 SRR id(s) from RNAseq to its tissue part, WGS SRR id and individual id}
\usage{
get_correspond_sample_info(inSRR, outdf)
}
\arguments{
\item{inSRR}{one or a list of GTEx RNAseq SRR id}

\item{outdf}{a dataframe idicating the mapping result}
}
\value{
a dataframe containing columns:individual,RNAseq_run,RNAseq_sample,RNAseq_tissue,WGS_run,WGS_sample,WGS_tissue
}
\description{
Function used to map GTEx v7 SRR id(s) from RNAseq to its tissue part, WGS SRR id and individual id
}
\examples{

inSRR:
c("SRR598300")

out_df:
  individual   RNAseq_run   RNAseq_sample   RNAseq_tissue   WGS_run   WGS_sample   WGS_tissue
  GTEX-PX3G   SRR598300   GTEX-PX3G-0006-SM-2I3E4   Whole Blood   SRR3478911   GTEX-PX3G-0004-SM-5JK53   Whole Blood

}
