% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sumabsdsg.R
\name{get_sumabsdsg}
\alias{get_sumabsdsg}
\title{For each individual, summarize loci infomation (DSGsum, DSGmean, DPsum, DPmean, absDSGsum, absDSGmean) by groups of loci defined by loci_list}
\usage{
get_sumabsdsg(mydir, loci_list)
}
\arguments{
\item{mydir}{dir containning .vcfs created by gangSTR}

\item{loci_list}{a list containing groups of loci}
}
\value{
dataframe with summarized information
example:

          sample eachdis DSGsum            DSGmean DPsum           DPmean absDSGsum absDSGmean      norm_dsg norm_absdsg
2 SRR2155717.vcf       1    334  0.878947368421053 16588 43.6526315789474       766  2.0157895  0.0201350374  0.04617796
3 SRR2155717.vcf       2    -70 -0.171990171990172 18435 45.2948402948403       490  1.2039312 -0.0037971250  0.02657988
4 SRR2155717.vcf       3     23 0.0484210526315789 21917 46.1410526315789       473  0.9957895  0.0010494137  0.02158142
5 SRR2155717.vcf       4     25 0.0389408099688473 28709 44.7180685358255       491  0.7647975  0.0008708071  0.01710265
6 SRR2155717.vcf       5     77  0.133913043478261 26145 45.4695652173913       629  1.0939130  0.0029451138  0.02405814
}
\description{
For each individual, summarize loci infomation (DSGsum, DSGmean, DPsum, DPmean, absDSGsum, absDSGmean) by groups of loci defined by loci_list
}
\examples{

loci_list example as defined by exon boundary distance:

[[1]]
[1] "chr1__6931916"    "chr1__10718732"   "chr1__25757550"   "chr1__31666544"   "chr1__32739601"

[[2]]
[1] "chr1__36032231"   "chr1__40255942"   "chr1__40713708"   "chr1__42094312"



}
