% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot.R
\name{lm_plot}
\alias{lm_plot}
\title{Title}
\usage{
lm_plot(Data)
}
\arguments{
\item{Data}{dataframe with two columns, x column and y column}
}
\value{
a pic with coef, r2, and pvalue
}
\description{
Title
}
\examples{

}
