% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot_covAgeGender.R
\name{lm_plot_covAgeGender}
\alias{lm_plot_covAgeGender}
\title{multivariate linear regression using gender and age as covariates}
\usage{
lm_plot_covAgeGender(myData)
}
\arguments{
\item{Data}{dataframe with FOUR columns, x,y,age,gender. formula: y~x+age+gender, exclude gender if gender are the same}
}
\value{
a pic with coef, adj.r.squared, and pvalue
}
\description{
multivariate linear regression using gender and age as covariates
}
\examples{

}
