% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_csiro.R
\name{scale_colour_csiro}
\alias{scale_colour_csiro}
\title{Title}
\usage{
scale_colour_csiro(..., palette = "core")
}
\arguments{
\item{...}{additional arguments passed to \code{\link[ggplot2]{discrete_scale}}}

\item{palette}{One of core, primary, secondary}
}
\description{
Title
}
