% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchWord.R
\name{searchWord}
\alias{searchWord}
\title{Search a string}
\usage{
searchWord(tree, string)
}
\arguments{
\item{tree}{an existing ternary search tree to search words in.}

\item{string}{a string of characters to search for.}
}
\value{
logical. TRUE if the string is in the tree, FALSE if not.
}
\description{
Searches for a string within a ternary search tree
}
\details{
Searches through the tree for the specified string. 
Returns TRUE if the string have been added to the tree and FALSE if not.
}
\examples{
fruitTree <- newTree(c("apple", "orange","apricot","cherry"))
searchWord(fruitTree, "apricot")
searchWord(fruitTree, "banana")
}

