% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_TCGA_id.R
\name{shorten_TCGA_id}
\alias{shorten_TCGA_id}
\title{used to conveniently cut TCGA sample IDs}
\usage{
shorten_TCGA_id(id_list, left_number)
}
\arguments{
\item{id_list}{individual list, e.g. TCGA-AAAA-BBBB}

\item{left_number}{how many parts (separated by '-' to keep, should not be longer than its original length )}
}
\value{
a shortened list
}
\description{
used to conveniently cut TCGA sample IDs
}
\examples{

}
