% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_somatic_vs_germline.R
\name{str_somatic_vs_germline}
\alias{str_somatic_vs_germline}
\title{used to analysis the str profile, compare them to the correspinding germline profile}
\usage{
str_somatic_vs_germline(RNAseq_GT, WGS_GT)
}
\arguments{
\item{RNAseq_GT}{result of tSTR::vcf2GT() and tSTR::get_correspond_sample_info()}

\item{WGS_GT}{result of tSTR::vcf2GT() and tSTR::get_correspond_sample_info()}
}
\value{
a dataframe include columns:loci T_gt_GT,TG1,TG2,T_REF,G_gt_GT,GG1,GG2,G_REF,dsg_modi,dsg_mode
}
\description{
used to analysis the str profile, compare them to the correspinding germline profile
}
\examples{

result dataframe

      loci    T_gt_GT  TG1    TG2      T_REF         G_gt_GT       GG1           GG2           G_REF            dsg_modi   dsg_mode
 1__10034928     1/1   gatg   gatg gatggatggatggatg     0/0 gatggatggatggatg gatggatggatggatg gatggatggatggatg       24 mutate_all
 1__10035556     1/1  aaaac  aaaac  aaaacaaaacaaaac     0/0  aaaacaaaacaaaac  aaaacaaaacaaaac  aaaacaaaacaaaac       20 mutate_all
 1__100405174     1/1 ttgttg ttgttg  ttgttgttgttgttg     0/0  ttgttgttgttgttg  ttgttgttgttgttg  ttgttgttgttgttg       18 mutate_all

}
