% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_csiro.R
\name{theme_csiro}
\alias{theme_csiro}
\title{CSIRO theme}
\usage{
theme_csiro(base_size = 12, base_family = "Calibri")
}
\arguments{
\item{base_size}{Base font size from plot}

\item{base_family}{Base font family}
}
\value{
A ggplot theme
}
\description{
Produce a ggplot \code{\link[ggplot2]{theme}} in line with the CSIRO Brand Identity Guidelines
}
