% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSTr-package.R
\docType{package}
\name{TSTr-package}
\alias{TSTr}
\alias{TSTr-package}
\title{\Sexpr{tools:::Rd_package_title("TSTr")}}
\description{
\Sexpr{tools:::Rd_package_description("TSTr")}
}
\details{
This package can be used to create a ternary search tree, more space efficient compared to standard prefix trees. Common applications for ternary search trees include spell-checking and auto-completion.
}
\author{
\Sexpr{tools:::Rd_package_author("TSTr")}

Maintainer:
\Sexpr{tools:::Rd_package_maintainer("TSTr")}
}
\references{
\url{https://en.wikipedia.org/wiki/Ternary_search_tree}
}
\seealso{
\code{\link{newTree}}
}
\keyword{package}

