% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2GT2dsg2DP.R
\name{vcf2GT}
\alias{vcf2GT}
\title{extract infomation of STR from gangSTR vcfs}
\usage{
vcf2GT(vcf_file)
}
\arguments{
\item{vcf_file}{the out file of gangSTR}
}
\value{
GT, a dataframe extracted from eachvcf
}
\description{
extract infomation of STR from gangSTR vcfs
}
\examples{
s=vcf2GT(vcf_file) \%>\% GT2maxlen()
s=vcf2GT(vcf_file) \%>\% GT2dsg()
s=vcf2DP(vcf_file)

GT
ChromKey,POS,Indiv,gt_GT,gt_GT_alleles,REF
1,14070,SRR656421Aligned.sortedByCoord.out,.,,cctccctccctc
1,16620,SRR656421Aligned.sortedByCoord.out,0/0,gctgctgctgct/gctgctgctgct,gctgctgctgct
1,22812,SRR656421Aligned.sortedByCoord.out,0/0,aggaaaggaa/aggaaaggaa,aggaaaggaa
1,26454,SRR656421Aligned.sortedByCoord.out,.,,gtgtgtgtgtgt

dsg:
ChromKey    POS SRR600493Aligned.sortedByCoord.out
    1  16620                                  0
    1  22812                                 10
    1 138589                                -12


maxlen:
    1  16620                                 12
    1  22812                                 15
    1 138589                                 12

DP:
 CHROM   POS   DP
 chr1 14070 <NA>
 chr1 16620    6
 chr1 22812    2


}
