% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSTr-package.R
\docType{data}
\name{XMIwords}
\alias{XMIwords}
\title{10001 most frequent English words}
\format{The format is: chr [1:10001] "the" "and" "that" "for" "you" "with"
"was" "this" "have" "but" "are" "not" "from" ...}
\description{
A character vector containing 10001 frequency-ordered English words. All words have 3 or more letters.
}
\examples{
data(XMIwords)
str(XMIwords)
}
\references{
Extracted from the English HC Corpora.
}
\keyword{datasets}

