% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSTr.r
\docType{package}
\name{TSTr}
\alias{TSTr}
\alias{TSTr-package}
\title{TSTr: Ternary Search Trees}
\description{
TSTr can be used to create a ternary search tree, more space efficient 
compared to standard prefix trees. Common applications for ternary search 
trees include spell-checking and auto-completion.
}
\details{
To learn more about TSTr, start with the vignettes:
\code{browseVignettes(package = "TSTr")}
}
\references{
\url{https://en.wikipedia.org/wiki/Ternary_search_tree}
}
\seealso{
\code{\link{newTree}}
}
\keyword{package}


